+++ src/Renderer/mesh.cpp
@@ -1,11 +1,11 @@
 ﻿#include "mesh.h"
 #include "../Core/utilities.h"
 
-Mesh::Mesh(float *vertexData, size_t vCount, uint16 *indexData, size_t iCount, Material material, AABB boundingBox)
+Mesh::Mesh(float *vertexData, size_t vCount, uint32 *indexData, size_t iCount, Material material, AABB boundingBox)
     : material(material), boundingBox(boundingBox)
 {
     ASSERT(vertexData && indexData);
-    auto iBuffer = std::make_shared<IndexBuffer>(iCount*sizeof(uint16), indexData);
+    auto iBuffer = std::make_shared<IndexBuffer>(iCount*sizeof(uint32), indexData);
 
         BufferLayout layout = {
             {BufferElement::Float3, "a_Position" },
